+++ Assets/Scripts/Player/UpdateStats.cs
@@ -18,7 +18,7 @@ public class UpdateStats : MonoBehaviour
     void Update()
     {
         player1CoinsText.text = Mathf.FloorToInt(PlayerDataManager.THIS.GetPlayer(0).GetCoins()).ToString();
-        player1PowerText.text = Mathf.FloorToInt(PlayerDataManager.THIS.GetPlayer(0).GetPower()).ToString();
+        player1PowerText.text = Mathf.FloorToInt((int)PlayerDataManager.THIS.GetPlayer(0).GetPower()).ToString();
         player2CoinsText.text = Mathf.FloorToInt(PlayerDataManager.THIS.GetPlayer(1).GetCoins()).ToString();
         player2PowerText.text = Mathf.FloorToInt(PlayerDataManager.THIS.GetPlayer(1).GetPower()).ToString();
     }
